feat: git infos are now cached by path
Git infos such as branch name and rebase status are stored in a global
variable. Previously identifiers of the global variable depended on the
pid. They now depend on the path to the folder. We cache the path to
make it identifier safe.
This achieve multiple things:
- If the current folder was previously resolved then we already have the
informations in a global variable. This allows to instantly show a
prompt with git infos when the path was already hit before.
- A lot less global variables
- Simplify the code